refactor(payment): updated paypal-commerce-integration package files structure
