+++ packages/paypal-commerce-integration/src/PayPalCommerceCredit/PayPalCommerceCreditPaymentMethod.test.tsx
@@ -1,36 +1,30 @@
 import { createCheckoutService, LanguageService } from '@bigcommerce/checkout-sdk';
-
-import { getPaymentFormServiceMock } from '@bigcommerce/checkout/test-mocks';
-
 import { render } from '@testing-library/react';
-import '@testing-library/jest-dom';
+import '@testing-library/jest-dom'
 import React from 'react';
 
 import { PaymentFormService } from '@bigcommerce/checkout/payment-integration-api';
 
-import { getPayPalCommerceAPMsMethod } from './mocks/paymentMethods.mock';
-import PayPalCommerceAPMsPaymentMethod from './PayPalCommerceAPMsPaymentMethod';
+import { getPayPalCommerceCreditMethod } from '../mocks/paymentMethods.mock';
+import PayPalCommerceCreditPaymentMethod from './PayPalCommerceCreditPaymentMethod';
 
-describe('PayPalCommerceAPMsPaymentMethod', () => {
+
+describe('PayPalCommerceCreditPaymentMethod', () => {
     const checkoutService = createCheckoutService();
     const checkoutState = checkoutService.getState();
-    const paymentForm: PaymentFormService = getPaymentFormServiceMock();
-
     const props = {
-        method: getPayPalCommerceAPMsMethod(),
+        method: getPayPalCommerceCreditMethod(),
         checkoutService,
         checkoutState,
-        paymentForm,
+        // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+        paymentForm: jest.fn() as unknown as PaymentFormService,
+        // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
         language: { translate: jest.fn() } as unknown as LanguageService,
         onUnhandledError: jest.fn(),
     };
 
-    beforeEach(() => {
-        jest.spyOn(checkoutState.data, 'isPaymentDataRequired').mockReturnValue(true);
-    });
-
     it('renders component with provided props', () => {
-        const { container } = render(<PayPalCommerceAPMsPaymentMethod {...props} />);
+        const { container } = render(<PayPalCommerceCreditPaymentMethod {...props} />);
 
         expect(container).toMatchSnapshot();
     });
@@ -43,12 +37,12 @@ describe('PayPalCommerceAPMsPaymentMethod', () => {
                 ...checkoutState,
                 data: {
                     ...checkoutState.data,
-                    isPaymentDataRequired: jest.fn().mockReturnValue(false),
-                },
+                    isPaymentDataRequired: jest.fn().mockReturnValue(false)
+                }
             },
-            children: mockChild,
-        };
-        const { container } = render(<PayPalCommerceAPMsPaymentMethod {...localProps} />);
+            children: mockChild
+        }
+        const { container } = render(<PayPalCommerceCreditPaymentMethod {...localProps} />);
 
         expect(container).toBeEmptyDOMElement();
     });
