+++ packages/paypal-commerce-integration/src/index.ts
@@ -1,6 +1,6 @@
-export { default as PayPalCommerceAPMsPaymentMethod } from './PayPalCommerceAPMsPaymentMethod';
-export { default as PayPalCommerceCreditPaymentMethod } from './PayPalCommerceCreditPaymentMethod';
+export { default as PayPalCommerceAPMsPaymentMethod } from './PayPalCommerceAPMs/PayPalCommerceAPMsPaymentMethod';
+export { default as PayPalCommerceCreditPaymentMethod } from './PayPalCommerceCredit/PayPalCommerceCreditPaymentMethod';
 export { default as PayPalCommerceFastlanePaymentMethod } from './PayPalCommerceFastlane';
-export { default as PayPalCommercePaymentMethod } from './PayPalCommercePaymentMethod';
-export { default as PayPalCommerceVenmoPaymentMethod } from './PayPalCommerceVenmoPaymentMethod';
-export { default as PaypalCommerceRatePayPaymentMethod } from './PaypalCommerceRatePayPaymentMethod';
+export { default as PayPalCommercePaymentMethod } from './PayPalCommerce/PayPalCommercePaymentMethod';
+export { default as PayPalCommerceVenmoPaymentMethod } from './PayPalCommerceVenmo/PayPalCommerceVenmoPaymentMethod';
+export { default as PaypalCommerceRatePayPaymentMethod } from './PayPalCommerceRatepay/PaypalCommerceRatePayPaymentMethod';
